Working on Jonah's requested changes
