+++ server/tests/ballot_polling/test_minerva_ballot_polling.py
@@ -5,6 +5,18 @@ from ..helpers import *  # pylint: disable=wildcard-import
 
 
 @pytest.fixture
+def election_id(client: FlaskClient, org_id: str, request):
+    set_logged_in_user(client, UserType.AUDIT_ADMIN, DEFAULT_AA_EMAIL)
+    return create_election(
+        client,
+        audit_name=f"Test Audit {request.node.name}",
+        audit_type=AuditType.BALLOT_POLLING,
+        audit_math_type=AuditMathType.MINERVA,
+        organization_id=org_id,
+    )
+
+
+@pytest.fixture
 def election_settings(client: FlaskClient, election_id: str):
     settings = {
         "electionName": "Test Election",
@@ -15,30 +27,6 @@ def election_settings(client: FlaskClient, election_id: str):
     }
     rv = put_json(client, f"/api/election/{election_id}/settings", settings)
     assert_ok(rv)
-
-
-@pytest.fixture
-def contest_ids(
-    client: FlaskClient, election_id: str, jurisdiction_ids: List[str]
-) -> List[str]:
-    contests = [
-        {
-            "id": str(uuid.uuid4()),
-            "name": "Contest 1",
-            "isTargeted": True,
-            "choices": [
-                {"id": str(uuid.uuid4()), "name": "candidate 1", "numVotes": 600,},
-                {"id": str(uuid.uuid4()), "name": "candidate 2", "numVotes": 400,},
-            ],
-            "totalBallotsCast": 1000,
-            "numWinners": 1,
-            "votesAllowed": 1,
-            "jurisdictionIds": jurisdiction_ids,
-        },
-    ]
-    rv = put_json(client, f"/api/election/{election_id}/contest", contests)
-    assert_ok(rv)
-    return [str(c["id"]) for c in contests]
 
 
 def test_minerva_ballot_polling_one_round(
@@ -78,6 +66,11 @@ def test_minerva_ballot_polling_one_round(
             contests[0]["choices"][0]["id"]: 30,
             contests[0]["choices"][1]["id"]: 20,
         },
+        contests[1]["id"]: {
+            contests[1]["choices"][0]["id"]: 20,
+            contests[1]["choices"][1]["id"]: 30,
+            contests[1]["choices"][2]["id"]: 10,
+        },
     }
 
     rv = put_json(
@@ -110,6 +103,11 @@ def test_minerva_ballot_polling_one_round(
         contests[0]["id"]: {
             contests[0]["choices"][0]["id"]: 20,
             contests[0]["choices"][1]["id"]: 10,
+        },
+        contests[1]["id"]: {
+            contests[1]["choices"][0]["id"]: 10,
+            contests[1]["choices"][1]["id"]: 15,
+            contests[1]["choices"][2]["id"]: 10,
         },
     }
 
@@ -182,6 +180,11 @@ def test_minerva_ballot_polling_two_rounds(
             contests[0]["choices"][0]["id"]: 20,
             contests[0]["choices"][1]["id"]: 20,
         },
+        contests[1]["id"]: {
+            contests[1]["choices"][0]["id"]: 10,
+            contests[1]["choices"][1]["id"]: 15,
+            contests[1]["choices"][2]["id"]: 10,
+        },
     }
 
     rv = put_json(
@@ -197,13 +200,6 @@ def test_minerva_ballot_polling_two_rounds(
     assert rv.status_code == 200
     assert json.loads(rv.data) == jurisdiction_1_results
 
-    # Round shouldn't be over yet, since we haven't recorded results for all jurisdictions with sampled ballots
-    rv = client.get(
-        f"/api/election/{election_id}/jurisdiction/{jurisdiction_ids[1]}/round"
-    )
-    rounds = json.loads(rv.data)["rounds"]
-    assert rounds[0]["endedAt"] is None
-
     rv = post_json(
         client,
         f"/api/election/{election_id}/jurisdiction/{jurisdiction_ids[1]}/round/{round_id}/audit-board",
@@ -214,6 +210,11 @@ def test_minerva_ballot_polling_two_rounds(
         contests[0]["id"]: {
             contests[0]["choices"][0]["id"]: 10,
             contests[0]["choices"][1]["id"]: 10,
+        },
+        contests[1]["id"]: {
+            contests[1]["choices"][0]["id"]: 10,
+            contests[1]["choices"][1]["id"]: 15,
+            contests[1]["choices"][2]["id"]: 10,
         },
     }
 
@@ -268,6 +269,11 @@ def test_minerva_ballot_polling_two_rounds(
             contests[0]["choices"][0]["id"]: 50,
             contests[0]["choices"][1]["id"]: 0,
         },
+        contests[1]["id"]: {
+            contests[1]["choices"][0]["id"]: 20,
+            contests[1]["choices"][1]["id"]: 30,
+            contests[1]["choices"][2]["id"]: 10,
+        },
     }
 
     rv = put_json(
@@ -283,13 +289,6 @@ def test_minerva_ballot_polling_two_rounds(
     assert rv.status_code == 200
     assert json.loads(rv.data) == jurisdiction_1_results
 
-    # Round shouldn't be over yet, since we haven't recorded results for all jurisdictions with sampled ballots
-    rv = client.get(
-        f"/api/election/{election_id}/jurisdiction/{jurisdiction_ids[1]}/round"
-    )
-    rounds = json.loads(rv.data)["rounds"]
-    assert rounds[1]["endedAt"] is None
-
     rv = post_json(
         client,
         f"/api/election/{election_id}/jurisdiction/{jurisdiction_ids[1]}/round/{round_2_id}/audit-board",
@@ -300,6 +299,11 @@ def test_minerva_ballot_polling_two_rounds(
         contests[0]["id"]: {
             contests[0]["choices"][0]["id"]: 30,
             contests[0]["choices"][1]["id"]: 0,
+        },
+        contests[1]["id"]: {
+            contests[1]["choices"][0]["id"]: 20,
+            contests[1]["choices"][1]["id"]: 30,
+            contests[1]["choices"][2]["id"]: 10,
         },
     }
